refactor: NavBar -> components/commons로 이동, MainLayout -> components/layouts로 이동
